+++ java/shared/java/org/signal/libsignal/internal/NativeTesting.kt
@@ -179,6 +179,9 @@ public object NativeTesting {
   ): String
 
   @JvmStatic
+  public external fun TESTING_DeleteAccountTests(): Array<Object>
+
+  @JvmStatic
   public external fun TESTING_DeleteBackupMediaTests(): Array<Object>
 
   @JvmStatic