+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@ v0.101.3
 
 - All libsignal-net-chat APIs that have both WebSocket and gRPC implementations will default to gRPC. This was already the case in practice for all three Signal clients due to remote configs *except* for `UnauthBackupsService.getUploadForm` and `getMediaUploadForm` (controlled by `grpc.BackupsAnonymousGetUploadForm`). This does not affect APIs that only have one implementation (a few older ones that are WebSocket-only still, and all the new ones that are gRPC-only).
 - New typed APIs:
+  - `AuthAccountsService.deleteAccount()`
   - `checkSVRCredentials()`
   - `getCurrencyConversions()`
   - `AuthKeysService.getPreKeyCount()`